move ir rewriting to lower module
