+++ lorax/src/lib.rs
@@ -1,10 +1,10 @@
 mod ir;
 mod link;
+mod lower;
 mod pool;
 mod rewrite;
 
-pub use ir::{
-    Block, Constant, OpBuilder, OpResult, Operation, RewritingCtx, Value, Var, rewrite_block,
-};
+pub use ir::{Block, Constant, OpBuilder, OpResult, Operation, Value, Var};
+pub use lower::{RewritingCtx, rewrite_block};
 pub use pool::{Pool, Ptr};
 pub use rewrite::{RewriteRule, RewriteRuleSet};
